Move worker IPC methods into IPC directory
